pyproject.toml
[project]
name = "spotify-img-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "pillow>=11.3.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.54.1",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
# TODO: Does this need other stuff for non-windows
torch = [{ index = "pytorch-cu128", marker = "sys_platform == 'win32'" }]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]
